+++ src/modules/ProgrammeManager/AppShell.jsx
@@ -1,6 +1,7 @@
 import { useRef } from 'react';
 import { TaskProvider } from './context/TaskContext';
 import { ViewProvider } from './context/ViewContext';
+import { CalendarProvider } from './context/CalendarContext';
 import RibbonTabs from './components/RibbonTabs/RibbonTabs';
 import SidebarTree from './components/SidebarTree';
 import TaskGrid from './components/TaskGrid';
@@ -57,8 +58,10 @@ function AppShellContent() {
 
 export default function AppShell() {
   return (
-    <ViewProvider>
-      <AppShellContent />
-    </ViewProvider>
+    <CalendarProvider>
+      <ViewProvider>
+        <AppShellContent />
+      </ViewProvider>
+    </CalendarProvider>
   );
 }